Sentiment analysis updates
