+++ packages/plugin-dkg/src/index.ts
@@ -1,6 +1,7 @@
 import { Plugin } from "@elizaos/core";
 
 import { dkgInsert } from "./actions/dkgInsert.ts";
+import { dkgAnalyzeSentiment } from "./actions/dkgAnalyzeSentiment.ts";
 
 import { graphSearch } from "./providers/graphSearch.ts";
 
@@ -14,7 +15,7 @@ export const dkgPlugin: Plugin = {
     name: "dkg",
     description:
         "Agent DKG which allows you to store memories on the OriginTrail Decentralized Knowledge Graph",
-    actions: [dkgInsert],
+    actions: [dkgInsert, dkgAnalyzeSentiment],
     providers: [graphSearch],
     evaluators: [sentimentAnalysisEvaluator],
 };
